requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
bidict==0.22.1
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
h11==0.14.0
idna==3.6
multidict==6.0.4
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.0.0
urllib3==2.1.0
websocket-client==1.7.0
wsproto==1.2.0
yarl==1.9.3
